+++ src/main/java/net/contargo/osrmproxy/config/MockConfiguration.java
@@ -21,22 +21,41 @@ public class MockConfiguration {
 
     @Profile("dev")
     @Bean(destroyMethod = "destroy")
-    public MockConfiguration createTwoMapsServerMocks(ProfilesProperties profilesProperties) {
+    public MockConfiguration createTwoMapsServerMocks(ProxyProperties proxyProperties) {
 
-        Integer port = profilesProperties.getDestinationOverLimit().getPort();
+        Integer port = proxyProperties.getProfiles().getLimit().get("driving").getAllDestinations().get(0).getPort();
         serverMock = new ServerMock(port);
 
-        double limitInMeters = profilesProperties.getLimitInMeters();
-
-        serverMock.configMock(getServerName(profilesProperties.getDestinationUnderLimit()),
-            "You arrived here because the aerial distance of your request is under the limit of " + limitInMeters);
-        serverMock.configMock(getServerName(profilesProperties.getDestinationOverLimit()),
-            "You arrived here because the aerial distance of your request is over the limit of " + limitInMeters);
-        serverMock.configMock(getServerName(profilesProperties.getFallbackDestination()),
-            "You arrived here because the request wasn't a routing request or "
-            + "there was an error while determine the aerial distance for the routing request");
+        proxyProperties.getProfiles().getLimit().values().forEach(this::configLimitMock);
+        proxyProperties.getProfiles().getSimple().values().forEach(this::configSimpleMock);
 
         return this;
+    }
+
+
+    private void configSimpleMock(SimpleProfileProperties properties) {
+
+        String serverName = getServerName(properties.getDestination());
+        serverMock.configMock(serverName, "You arrived here because of the specified profile");
+    }
+
+
+    private void configLimitMock(LimitProfileProperties properties) {
+
+        double limitInMeters = properties.getLimitInMeters();
+        LimitProfileProperties.Destinations destinations = properties.getDestinations();
+
+        serverMock.configMock(getServerName(destinations.getUnderLimit()),
+            "You arrived here because the aerial distance of your request is under the limit of "
+            + limitInMeters);
+
+        serverMock.configMock(getServerName(destinations.getOverLimit()),
+            "You arrived here because the aerial distance of your request is over the limit of "
+            + limitInMeters);
+
+        serverMock.configMock(getServerName(destinations.getFallback()),
+            "You arrived here because the request wasn't a routing request or "
+            + "there was an error while determine the aerial distance for the routing request");
     }