+++ src/main/java/net/contargo/osrmproxy/health/OsrmHealthIndicator.java
@@ -1,6 +1,8 @@
 package net.contargo.osrmproxy.health;
 
-import net.contargo.osrmproxy.config.ProfilesProperties;
+import net.contargo.osrmproxy.config.LimitProfileProperties;
+import net.contargo.osrmproxy.config.ProxyProperties;
+import net.contargo.osrmproxy.config.SimpleProfileProperties;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,25 +23,46 @@ import java.lang.invoke.MethodHandles;
 
 import java.net.URL;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.toList;
 
 
 /**
  * @author  Sandra Thieme - thieme@synyx.de
+ * @author  Ben Antony - antony@synyx.de
  */
 @Component
 public class OsrmHealthIndicator implements HealthIndicator {
 
     private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
-    private final ProfilesProperties profilesProperties;
+    private final List<URL> osrmInstances;
     private final RestTemplate restTemplate;
+    private final ProxyProperties.HealthCoordinates coordinates;
 
-    public OsrmHealthIndicator(ProfilesProperties profilesProperties, RestTemplate restTemplate) {
+    public OsrmHealthIndicator(ProxyProperties properties, RestTemplate restTemplate) {
 
-        this.profilesProperties = profilesProperties;
         this.restTemplate = restTemplate;
+        this.coordinates = properties.getHealthCoordinates();
+
+        ProxyProperties.Profiles profiles = properties.getProfiles();
+
+        List<URL> instances = profiles.getSimple()
+                .values()
+                .stream()
+                .map(SimpleProfileProperties::getDestination)
+                .collect(toList());
+
+        instances.addAll(profiles.getLimit()
+            .values()
+            .stream()
+            .map(LimitProfileProperties::getAllDestinations)
+            .flatMap(List::stream)
+            .collect(toList()));
+
+        this.osrmInstances = instances.stream().distinct().collect(toList());
     }
 
     @Override
@@ -47,10 +70,7 @@ public class OsrmHealthIndicator implements HealthIndicator {
 
         Health.Builder builder = new Health.Builder(Status.UP);
 
-        Stream.of(profilesProperties.getDestinationOverLimit(), profilesProperties.getDestinationUnderLimit(),
-                profilesProperties.getFallbackDestination())
-            .filter(this::isDown)
-            .forEach(u -> builder.down());
+        osrmInstances.stream().filter(this::isDown).forEach(u -> builder.down());
 
         return builder.build();
     }
@@ -58,10 +78,9 @@ public class OsrmHealthIndicator implements HealthIndicator {
 
     private boolean isDown(URL url) {
 
-        ProfilesProperties.HealthCoordinates c = profilesProperties.getHealthCoordinates();
-
-        String request = String.format(Locale.ENGLISH, "%s/route/v1/driving/%f,%f;%f,%f", url, c.getStartLongitude(),
-                c.getStartLatitude(), c.getEndLongitude(), c.getEndLatitude());
+        String request = String.format(Locale.ENGLISH, "%s/route/v1/driving/%f,%f;%f,%f", url,
+                coordinates.getStartLongitude(), coordinates.getStartLatitude(), coordinates.getEndLongitude(),
+                coordinates.getEndLatitude());
 
         LOG.debug("Health checking {}", request);