+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2-SNAPSHOT
+
+- Add multiple profile support:
+    - Two proxy types `simple` and `limit`
+    - **BREAKING CHANGES** the configuration changed, see [README](https://github.com/Contargo/osrm-profile-proxy/blob/master/README.md)
+
 ## 1.1
 
 - Add health check